lo que sea

+++ src/api/domain/adoption_process/controller.py
@@ -6,23 +6,30 @@ def get_all_adoption_processes():
     return Repository.get_all_adoption_processes()
 
 
-def create_adoption_process(new_process):
-    return Repository.create_adoption_process(new_process['user_id'], new_process['pet_id'], new_process['description'], new_process['status'])
+def create_adoption_process(new_process, status):
+    if pet_id is None:
+        return {"msg": "Animal no encontrado", "error": True, "status": 404 }
+    return Repository.create_adoption_process(new_process['user_id'], new_process['pet_id'], new_process['description'], status)
 
 
 def delete_adoption_process(adoption_process_id):
+    if user['id'] is None or user['id'] != volunteer['user_id'] :
+        return {"msg": "No tiene el acceso permitido", "error": True, "status": 400 }
+    
     if adoption_process_id is None:
-        return {'error': 'Proceso ID no indicado'}, 400
+        return {'error': 'ID Proceso de adopción no encontrado'}, 400
     adoption_process = Adoption_process.query.get(adoption_process_id)
     if not adoption_process:
-        return {'error': 'Proceso no encontrado'}, 404
+        return {'error': 'Proceso de adopción no encontrado'}, 404
     Repository.delete_adoption_process(adoption_process_id)
     return {'msg': 'Proceso de adopción borrado satisfactoriamente'}, 204
 
 def update_adoption_process(adoption_process_id, data):
+    if user['rol_id'] is None or user['rol_id'] != "ADMIN":
+         return {"msg": "No tiene el acceso permitido", "error": True, "status": 400 }
     adoption_process = Adoption_process.query.get(adoption_process_id)
     if not adoption_process:
-        return {'error': 'Proceso de adopción no encontrado'}, 404
+        return {"msg": "Proceso de adopción no encontrado", "error": True, "status": 400 }
     Repository.update_adoption_process(data, adoption_process_id)
     return {'msg': 'Proceso de adopción actualizado correctamente'}, 200
